Refactor: const.js 파일 위치 변경
/src/components/carouselLibR -> /src
